Change shape type to usize slice
[skip ci]

+++ src/tensors.rs
@@ -246,7 +246,7 @@ extern {
     ) -> Variable;
 
     /// Creates a Tensor with all elements set to 0.
-    pub fn zeroes(shape: &[u32], dtype: Option<DType>) -> Tensor;
+    pub fn zeroes(shape: &[usize], dtype: Option<DType>) -> Tensor;
 
     /// Creates a Tensor with all elements set to 0 with the same shape as the given tensor.
     pub fn zeroes_like(input: &JsValue) -> Tensor;
